Redesign menu logic
New menu logic uses cursor position to write the highlighted option, so instead of writing the whole menu
every time you go up or down, now it just moves the cursor to the position and writes the option highlighted
(while de-highlighting the previous option). This also allows for the menu options to be anywhere (they don't
have to start at first row) and because of that you can have the exercise solution stay on the screen after it
is done (because the screen doesn't refresh).